+++ src/main/java/controller/JanggiController.java
@@ -6,8 +6,8 @@ import domain.board.BoardFactory;
 import domain.board.Column;
 import domain.board.Position;
 import domain.board.Row;
-import domain.gameState.BlueTurn;
 import domain.piece.PieceType;
+import domain.state.BlueTurn;
 import dto.MoveCommandDTO;
 import view.InputView;
 import view.OutputView;
