+++ src/test/java/domain/state/RedTurnTest.java
@@ -1,4 +1,4 @@
-package domain.gameState;
+package domain.state;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -11,31 +11,31 @@ import domain.board.Row;
 import domain.piece.PieceType;
 import org.junit.jupiter.api.Test;
 
-class BlueTurnTest {
+class RedTurnTest {
     @Test
-    void 파란팀_차례일때_파란색_기물을_움직일수_있다() {
+    void 빨간팀_차례일때_빨간색_기물을_움직일수_있다() {
         BoardFactory boardFactory = new BoardFactory();
         Board board = boardFactory.createBoard();
-        State blueTurn = new BlueTurn(board);
+        State RedTurn = new RedTurn(board);
 
-        Position source = new Position(Row.ZERO, Column.ONE);
-        Position target = new Position(Row.EIGHT, Column.ONE);
+        Position source = new Position(Row.ONE, Column.ONE);
+        Position target = new Position(Row.THREE, Column.ONE);
 
-        assertThatCode(() -> blueTurn.movePiece(PieceType.CHARIOT, source, target))
+        assertThatCode(() -> RedTurn.movePiece(PieceType.CHARIOT, source, target))
                 .doesNotThrowAnyException();
 
     }
 
     @Test
-    void 파란팀_차례일때_기물이_없는_위치에서_움직일수_없다() {
+    void 빨간팀_차례일때_기물이_없는_위치에서_움직일수_없다() {
         BoardFactory boardFactory = new BoardFactory();
         Board board = boardFactory.createBoard();
-        State blueTurn = new BlueTurn(board);
+        State RedTurn = new RedTurn(board);
 
         Position source = new Position(Row.TWO, Column.ONE);
         Position target = new Position(Row.THREE, Column.ONE);
 
-        assertThatThrownBy(() -> blueTurn.movePiece(PieceType.EMPTY, source, target))
+        assertThatThrownBy(() -> RedTurn.movePiece(PieceType.EMPTY, source, target))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 }
